refactor events, improve UX, clean up float literals and fix bugs
refactor(events): improve UX, clean up float literals, and fix critical bugs

- Refactored all event classes to improve user experience and clarify code logic.
- Standardized float literals: changed uppercase suffixes (e.g., `1.0F`) to lowercase (`1.0f`) for consistency with Java conventions.
- Removed unnecessary explicit casts; e.g., replaced `(float) Settings.WIDTH / 2` with `Settings.WIDTH / 2.0f` for cleaner and more accurate floating-point division.
- Replaced `@see` Javadoc tags with `{@link}` for better IDE support and inline readability.
- Fixed a bug in the Spark Path event where cards were being upgraded even when the upgrade option was not selected.
- Reworked several relic effects to make them more viable and less underwhelming.
- Fixed an issue in the Abyss event that incorrectly granted Strength.
- Replaced raw Iterator loops with enhanced for-loops for better readability
  and type safety in NobleCardHelper.
- Removed lots of silly, redundant, obsolete, and self-explanatory comments to reduce noise
  and improve code clarity.
- Removed repeat plugins in pom.xml
- Improved user experience when interacting with the Noble Phantasm card group.

+++ src/main/java/fgo/cards/noblecards/EnferChateaudIf.java
@@ -36,6 +36,6 @@ public class EnferChateaudIf extends AbsNoblePhantasmCard {
 
 
     private void spawnRelicWingBoots() {
-        AbstractDungeon.getCurrRoom().spawnRelicAndObtain((float)(Settings.WIDTH / 2), (float)(Settings.HEIGHT / 2), new WingBoots());
+        AbstractDungeon.getCurrRoom().spawnRelicAndObtain(Settings.WIDTH / 2.0f, Settings.HEIGHT / 2.0f, new WingBoots());
     }
 }
